+++ projects/igniteui-angular-i18n/src/i18n/ZH-HANT/query-builder-resources.ts
@@ -42,7 +42,6 @@ export const QueryBuilderResourceStringsZHHANT = {
     igx_query_builder_filter_true: 'True',
     igx_query_builder_filter_false: 'False',
     igx_query_builder_filter_all: '全部',
-    igx_query_builder_title: 'Query Builder',
     igx_query_builder_from_label: '從',
     igx_query_builder_select_label: '選擇',
     igx_query_builder_where_label: '在哪裡',