+++ projects/igniteui-angular/migrations/update-19_1_0/index.spec.ts
@@ -75,10 +75,10 @@ describe(`Update to ${version}`, () => {
         );
     });
 
-    it('should remove Query Builder Header deprecated property `showLegend`', async () => {
+    it('should remove Query Builder Header deprecated `showLegend` and `resourceStrings` properties', async () => {
         appTree.create(`/testSrc/appPrefix/component/test.component.html`,
         `
-        <igx-query-builder-header [showLegend]="false"></igx-query-builder-header>
+        <igx-query-builder-header [showLegend]="false" [resourceStrings]="resourceStrings"></igx-query-builder-header>
         `
         );
 
@@ -87,6 +87,33 @@ describe(`Update to ${version}`, () => {
         expect(tree.readContent('/testSrc/appPrefix/component/test.component.html')).toEqual(
         `
         <igx-query-builder-header></igx-query-builder-header>
+        `
+        );
+    });
+
+    it('should remove igx_query_builder_title from resources ', async () => {
+        appTree.create(`/testSrc/appPrefix/component/test.component.ts`,
+        `
+        export class TestComponent {
+            public resourceStrings = {
+                igx_query_builder_add_condition: '+ Add Condition',
+                igx_query_builder_title: 'Query Builder Title',
+                igx_query_builder_add_group: '+ Add Group',
+            };
+        }
+        `
+        );
+
+        const tree = await schematicRunner.runSchematic(migrationName, { shouldInvokeLS: false }, appTree);
+
+        expect(tree.readContent('/testSrc/appPrefix/component/test.component.ts')).toEqual(
+        `
+        export class TestComponent {
+            public resourceStrings = {
+                igx_query_builder_add_condition: '+ Add Condition',
+                igx_query_builder_add_group: '+ Add Group',
+            };
+        }
         `
         );
     });